bot/pyproject.toml
[tool.poetry]
name = "sparrow_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {extras = ["voice"], version = "^1.4.1"}
aiohttp = "^3.6.2"
uvloop = "^0.14.0"
python-dotenv = "^0.14.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
